Rename error that occurs when writing on a read
I originally named this `StatementInvalid` because that's what we do in
GitHub, but `@tenderlove` pointed out that this means apps can't test
for or explitly rescue this error. `StatementInvalid` is pretty broad so
I've renamed this to `InvalidWriteOnRead`.

I also changed the `while_preventing_writes` block to reset to false -
we always want to turn this feature off at the end of the block.